refactor(session): reorganize flat files into domain-based directory structure
