+++ packages/session/src/index.ts
@@ -1,6 +1,5 @@
 export { Bus, BusEvent } from "./bus";
-export { Storage, InMemoryStorage } from "./storage";
-export { FileStorageAdapter } from "./file-storage";
+export { Storage, InMemoryStorage, FileStorageAdapter } from "./storage";
 export { Session } from "./session";
 export { SessionStatus } from "./status";
 export { Snapshot, InMemorySnapshotProvider } from "./snapshot";
