+++ packages/session/src/session/info.ts
